requirements/desarrollo.txt
-r base.txt
django-debug-toolbar
pytest-django
